code simplification and refactoring
