+++ bundles/org.openhab.binding.bluetooth.generic/src/main/java/org/openhab/binding/bluetooth/generic/internal/GenericBluetoothHandlerFactory.java
@@ -21,11 +21,7 @@ import org.openhab.core.thing.ThingTypeUID;
 import org.openhab.core.thing.binding.BaseThingHandlerFactory;
 import org.openhab.core.thing.binding.ThingHandler;
 import org.openhab.core.thing.binding.ThingHandlerFactory;
-import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-import org.osgi.service.component.annotations.ReferenceCardinality;
-import org.sputnikdev.bluetooth.gattparser.BluetoothGattParser;
 
 /**
  * The {@link GenericBluetoothHandlerFactory} is responsible for creating things and thing
@@ -40,14 +36,6 @@ public class GenericBluetoothHandlerFactory extends BaseThingHandlerFactory {
     private static final Set<ThingTypeUID> SUPPORTED_THING_TYPES_UIDS = Set
             .of(GenericBindingConstants.THING_TYPE_GENERIC);
 
-    private final BluetoothGattParser gattParser;
-
-    @Activate
-    public GenericBluetoothHandlerFactory(
-            @Reference(cardinality = ReferenceCardinality.MANDATORY) GattParserFactory gattParserFactory) {
-        this.gattParser = gattParserFactory.getParser();
-    }
-
     @Override
     public boolean supportsThingType(ThingTypeUID thingTypeUID) {
         return SUPPORTED_THING_TYPES_UIDS.contains(thingTypeUID);
@@ -58,7 +46,7 @@ public class GenericBluetoothHandlerFactory extends BaseThingHandlerFactory {
         ThingTypeUID thingTypeUID = thing.getThingTypeUID();
 
         if (GenericBindingConstants.THING_TYPE_GENERIC.equals(thingTypeUID)) {
-            return new GenericBluetoothHandler(thing, gattParser);
+            return new GenericBluetoothHandler(thing);
         }
 
         return null;
